Exercicio 3 - Dia 2 - Bloco 12
